organize gatsby api related files
